# ai-spec/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 优化家 MVP API
  description: |
    债务优化决策引擎 - MVP 阶段接口规范
    
    设计规范遵循 Google API Design Guide (https://google.aip.dev)：
    - 资源导向设计（AIP-121）
    - 标准方法 List/Get/Create/Update/Delete（AIP-130~135）
    - 自定义方法使用 :action 语法（AIP-136）
    - 字段名 camelCase（AIP-140）
    - 分页使用 pageToken（AIP-158）
    - 部分更新使用 PATCH + updateMask（AIP-134）
  version: 1.0.0
  contact:
    name: YouHuaJia Team
servers:
  - url: /api/v1
    description: API Base Path
...
components:
  schemas:
...
    FinanceProfile:
      type: object
      properties:
        name:
          type: string
          description: "资源名称: users/{userId}/finance-profile"
          readOnly: true
        totalDebt:
          type: number
        debtCount:
          type: integer
        weightedApr:
          type: number
          description: 加权年化利率（百分比）
        monthlyPayment:
          type: number
        monthlyIncome:
          type: number
        debtIncomeRatio:
          type: number
        liquidityScore:
          type: number
        restructureScore:
          type: number
          description: 重组评分（0-100）
        riskLevel:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        scoreDetail:
          type: object
          description: 评分各维度明细
        lastCalculateTime:
          type: string
          format: date-time
          readOnly: true

    # ---------- OcrTask 资源 ----------
...
    Report:
      type: object
      properties:
        name:
          type: string
          description: "资源名称: reports/{reportId}"
          readOnly: true
        profileSnapshot:
          $ref: '#/components/schemas/FinanceProfile'
        priorityList:
          type: array
          items:
            type: object
            properties:
              rank:
                type: integer
              debt:
                type: string
                description: "债务资源名称: debts/{debtId}"
              creditor:
                type: string
              apr:
                type: number
              reason:
                type: string
        actionPlan:
          type: object
          description: 90天行动计划
        aiSummary:
          type: string
          description: AI 生成的文字建议
        riskWarnings:
          type: array
          items:
            type: string
        reportVersion:
          type: integer
        createTime:
          type: string
          format: date-time
          readOnly: true
    ListReportsResponse:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: '#/components/schemas/Report'
        nextPageToken:
          type: string

    # ---------- Engine（无状态计算工具）----------
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []

# ============================================================
# 路径定义（资源导向 + 标准方法 + 自定义方法）
# ============================================================
paths:
...
  /scoring-strategies/{segment}:
    get:
      tags: [scoring-strategies]
      summary: 获取指定分群的策略详情
      operationId: getStrategy
      parameters:
        - name: segment
          in: path
          required: true
          schema:
            type: string
            enum: [DEFAULT, HIGH_DEBT, MORTGAGE_HEAVY, YOUNG_BORROWER]
      responses:
        '200':
          description: 策略详情
        '404':
          description: 策略不存在
...
  /scoring-strategies:compare:
    post:
      tags: [scoring-strategies]
      summary: 对比两个策略对同一用户的评分差异
      operationId: compareStrategies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, segmentA, segmentB]
              properties:
                userId:
                  type: integer
                  format: int64
                segmentA:
                  type: string
                segmentB:
                  type: string
      responses:
        '200':
          description: 对比结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  strategyA:
                    type: object
                    properties:
                      segment:
                        type: string
                      strategyName:
                        type: string
                      version:
                        type: string
                      finalScore:
                        type: number
                      riskLevel:
                        type: string
                      recommendation:
                        type: string
                  strategyB:
                    type: object
                    properties:
                      segment:
                        type: string
                      strategyName:
                        type: string
                      version:
                        type: string
                      finalScore:
                        type: number
                      riskLevel:
                        type: string
                      recommendation:
                        type: string
                  scoreDelta:
                    type: number
                    description: 分数差值 (A - B)

  # ===== Reports 资源 =====
  /reports:
    get:
      tags: [reports]
      summary: List - 报告历史列表
      operationId: listReports
      parameters:
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
        - name: pageToken
          in: query
          schema:
            type: string
        - name: orderBy
          in: query
          schema:
            type: string
            default: "createTime desc"
      responses:
        '200':
          description: 报告列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListReportsResponse'
  /reports:generate:
    post:
      tags: [reports]
      summary: 生成优化报告（自定义方法，基于当前画像）
      operationId: generateReport
      responses:
        '200':
          description: 报告生成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          description: 画像不完整
  /reports/{reportId}:
    get:
      tags: [reports]
      summary: Get - 获取报告详情
      operationId: getReport
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 报告详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
  /reports/{reportId}:export:
    get:
      tags: [reports]
      summary: 导出报告为 PDF（自定义方法）
      operationId: exportReport
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PDF 文件流
          content:
            application/pdf:
              schema:
                type: string
                format: binary